feat: create User and your Account

+++ apps/api_banking/priv/repo/migrations/20210413203344_create_account_user.exs
@@ -5,7 +5,7 @@ defmodule ApiBanking.Repo.Migrations.CreateAccounts do
     create table(:accounts, primary_key: false) do
       add :id, :uuid, primary_key: true
       add :account_code, :string
-      add :balance, :float
+      add :balance, :integer
 
       add :user_id, references(:users, type: :uuid)
